backend/order_service/pyproject.toml
[project]
name = "order-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiocache>=0.12.3",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.14",
    "fastapi-mail>=1.5.0",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "orjson>=3.11.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "watchfiles>=1.1.0",
]
